feat(endcap-backlit-multitier): TASK-654 — promote _check_illumination and _base_model_from_str to AbstractPlanogramType
