+++ packages/ai-parrot-pipelines/src/parrot_pipelines/planogram/types/endcap_no_shelves_promotional.py
@@ -458,104 +458,14 @@ class EndcapNoShelvesPromotional(AbstractPlanogramType):
         planogram_description: Any,
         illum_zone_bbox: Optional[Any] = None,
     ) -> str:
-        """Check backlit illumination state using the illuminated zone crop.
+        """Delegate to :meth:`AbstractPlanogramType._check_illumination`.
 
-        Crops only the zone declared as backlit (via ``illum_zone_bbox``) so
-        the LLM focuses on the header panel alone, avoiding confusion from
-        ambient store lighting in surrounding areas.  Falls back to the full
-        endcap ROI when no zone bbox is provided.
-
-        Args:
-            img: The full input PIL image.
-            roi: Endcap detection with a bbox attribute (used as fallback crop).
-            planogram_description: Planogram description for brand context.
-            illum_zone_bbox: DetectionBox with pixel coordinates of the
-                illuminated zone (e.g. the header shelf).  When provided,
-                this crop is used instead of the full endcap ROI.
-
-        Returns:
-            ``'illumination_status: ON'`` or ``'illumination_status: OFF'``
+        Retained for backward compatibility; the implementation now lives in
+        the abstract base class so all planogram types can share it.
         """
-        iw, ih = img.size
-
-        if illum_zone_bbox is not None:
-            # Crop only the illuminated zone for a focused check.
-            x1 = max(0, int(illum_zone_bbox.x1))
-            y1 = max(0, int(illum_zone_bbox.y1))
-            x2 = min(iw, int(illum_zone_bbox.x2))
-            y2 = min(ih, int(illum_zone_bbox.y2))
-            roi_crop = img.crop((x1, y1, x2, y2))
-            self.logger.debug(
-                "Illumination check using zone crop (%d,%d,%d,%d)", x1, y1, x2, y2
-            )
-        elif roi is not None and hasattr(roi, "bbox"):
-            x1 = int(roi.bbox.x1 * iw)
-            y1 = int(roi.bbox.y1 * ih)
-            x2 = int(roi.bbox.x2 * iw)
-            y2 = int(roi.bbox.y2 * ih)
-            roi_crop = img.crop((x1, y1, x2, y2))
-        else:
-            roi_crop = img.copy()
-
-        roi_small = self.pipeline._downscale_image(roi_crop, max_side=800, quality=82)
-        brand = (getattr(planogram_description, "brand", "") or "").strip()
-        brand_hint = f" {brand}" if brand else ""
-
-        prompt = (
-            f"You are a retail display inspector evaluating a{brand_hint} backlit "
-            "lightbox panel (the kind that has fluorescent or LED tubes BEHIND the "
-            "graphic, making the sign glow from within).\n\n"
-            "This image shows ONLY the header panel crop. Analyze it carefully.\n\n"
-            "A backlit lightbox that is ON shows ALL of these signs:\n"
-            "  1. The panel surface has a UNIFORM, EVEN glow — the brightness is "
-            "consistent across the entire face of the sign, not just where the store "
-            "lights happen to hit it.\n"
-            "  2. The aluminum or silver frame around the panel appears BRIGHT or has "
-            "a HALO/GLOW along its inner edge.\n"
-            "  3. The colors in the graphic look VIVID and SATURATED — backlit prints "
-            "appear translucent and luminous, not opaque like a regular poster.\n"
-            "  4. The panel is DISTINCTLY BRIGHTER than non-illuminated surfaces "
-            "nearby (ceiling, walls, shelving).\n\n"
-            "A backlit lightbox that is OFF shows:\n"
-            "  1. The graphic is visible but ONLY because of ambient store ceiling "
-            "lights — NOT because the sign itself is emitting light.\n"
-            "  2. The panel looks like a regular PRINTED POSTER or VINYL PRINT — "
-            "opaque, matte, with no translucent glow.\n"
-            "  3. The frame is dull/matte metal with NO halo.\n"
-            "  4. Brightness across the panel surface is UNEVEN — brighter where "
-            "overhead lights hit, darker in corners.\n\n"
-            "CRITICAL: A well-lit store can make ANY sign look bright. That does NOT "
-            "mean the backlight is ON. Look for self-emission, even luminosity, and "
-            "frame glow — these only occur when the internal light source is active.\n\n"
-            "First, briefly state what you observe (1-2 sentences). "
-            "Then on a new line write EXACTLY one word: LIGHT_ON or LIGHT_OFF"
+        return await super()._check_illumination(
+            img, roi, planogram_description, illum_zone_bbox
         )
-
-        raw_answer = ""
-        try:
-            async with self.pipeline.roi_client as client:
-                msg = await client.ask_to_image(
-                    image=roi_small,
-                    prompt=prompt,
-                    model=GoogleModel.GEMINI_3_FLASH_PREVIEW,
-                    no_memory=True,
-                    max_tokens=128,
-                )
-            raw_answer = (msg.output or "").strip().upper()
-        except Exception as exc:
-            self.logger.warning(
-                "Illumination check failed: %s — defaulting to ON", exc
-            )
-
-        state = (
-            "illumination_status: OFF"
-            if "LIGHT_OFF" in raw_answer
-            else "illumination_status: ON"
-        )
-        self.logger.info(
-            "Endcap illumination check → answer=%r  state=%s", raw_answer, state
-        )
-        return state
 
     def check_planogram_compliance(
         self,
