+++ packages/ai-parrot-pipelines/src/parrot_pipelines/planogram/types/product_on_shelves.py
@@ -869,73 +869,6 @@ class ProductOnShelves(AbstractPlanogramType):
 
         return endcap_det, panel_det, brand_det, text_det, dets
 
-    def _base_model_from_str(
-        self, s: str, brand: str = None, patterns: Optional[List[str]] = None
-    ) -> str:
-        """Extract normalized base model from any text, supporting multiple brands.
-
-        If ``patterns`` is provided (from planogram_config.model_normalization_patterns),
-        those regex patterns are tried first and replace the generic defaults.
-        Each pattern's captured groups are joined with '-' to form the key.
-        Brand-specific and generic fallbacks only run when no configured patterns exist.
-        """
-        if not s:
-            return ""
-
-        t = s.lower().strip()
-        t = t.replace("\u2014", "-").replace("\u2013", "-").replace("_", "-")
-
-        # Configured patterns (from DB) -- replace generic defaults when present
-        if patterns:
-            for pat in patterns:
-                m = re.search(pat, t)
-                if m:
-                    groups = [g for g in m.groups() if g]
-                    if groups:
-                        return "-".join(groups)
-            return ""
-
-        # Brand-specific fallback (kept for backward compatibility)
-        if brand and brand.lower() == "epson":
-            m = re.search(r"(et)[- ]?(\d{4})", t)
-            if m:
-                return f"{m.group(1)}-{m.group(2)}"
-
-        elif brand and brand.lower() == "hisense":
-            if re.search(r"canvas[\s-]*tv", t):
-                return "canvas-tv"
-            if re.search(r"canvas", t):
-                return "canvas"
-            hisense_patterns = [
-                r"(\d*)(u\d+)([a-z]*)",
-                r"(u\d+)",
-            ]
-            for pattern in hisense_patterns:
-                m = re.search(pattern, t)
-                if m:
-                    if len(m.groups()) >= 2:
-                        size = m.group(1) if m.group(1) else ""
-                        series = m.group(2)
-                        variant = m.group(3) if len(m.groups()) > 2 and m.group(3) else ""
-                        return f"{size}{series}{variant}".lower()
-                    else:
-                        return m.group(1).lower()
-
-        # Generic default
-        generic_patterns = [
-            r"([a-z]+)[- ]?(\d{2,4})",
-            r"([a-z]\d+)",
-            r"(\d{4})",
-        ]
-        for pattern in generic_patterns:
-            m = re.search(pattern, t)
-            if m:
-                if len(m.groups()) >= 2:
-                    return f"{m.group(1)}-{m.group(2)}"
-                else:
-                    return m.group(1).lower()
-        return ""
-
     def _canonical_expected_key(
         self, sp: Any, brand: str, patterns: Optional[List[str]] = None
     ) -> Tuple[str, str]:
